Moved c_shared_lib to __init__.py and imported it in other files
