+++ source/tomopy/util/extern/__init__.py
@@ -49,35 +49,10 @@
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
-import sys
-import os.path
-import ctypes
-import logging
-
-logger = logging.getLogger(__name__)
-
-
-def c_shared_lib(lib_name, do_warn=True):
-    """Get the path and import the C-shared library."""
-    load_dll = ctypes.cdll.LoadLibrary
-    ext = '.so'
-    if sys.platform == 'darwin':
-        ext = '.dylib'
-    if os.name == 'nt':
-        ext = '.dll'
-        load_dll = ctypes.windll.LoadLibrary
-    base_path = os.path.abspath(os.path.dirname(
-                                os.path.dirname(__file__)))
-    sharedlib = os.path.join(base_path, 'sharedlibs', '%s%s' % (lib_name, ext))
-    if os.path.exists(sharedlib):
-        return load_dll(sharedlib)
-    # cannot find shared lib:
-    if do_warn is True:
-        logger.warning('OSError: The following shared lib is missing!\n{}'.format(
-                       sharedlib))
+from tomopy.util.extern.recon import *
+from tomopy.util.extern.accel import *
+from tomopy.util.extern.gridrec import *
+from tomopy.util.extern.prep import *
+from tomopy.util.extern.misc import *
 
 
-def MissingLibrary(function):
-    print(f"The {function} algorithm is unavailable."
-          " Check CMake logs to determine if TomoPy was"
-          " built with dependencies required by this algorithm.")
